Refactor AssertionList into the Model directory
